+++ CHANGELOG.md
@@ -55,6 +55,7 @@
     - Handle callback. (#183, #208)
 - Implement `claimBoost`:
     - Refund requester. (#198, #201)
+    - Seize workerpool and reward kitty. (#218)
 - Add dedicated Poco Boost accessors. (#206)
 - Refactor tests. (#177, #178, #179, #181, #188, #200, #209, #210, #211)
 - Add mocha to use test explorer. (#180)